+++ src/physique/data/UtilisateurServiceORM.java
@@ -24,5 +24,5 @@ public interface UtilisateurServiceORM {
     public List<Utilisateur> getAll();
     public List<Utilisateur> getAllByRange(int debut,int fin);
     public List<Technicien> getByLogin(String login);
-    public int count();
+    public Long count();
 }
